Spoon & 7th mission campaign
1. Изменена структура сохранений с целью уменьшения занимаемого места, теперь вместо папок actions & snapshots, с многими маленькими файликами внутри, есть несколько больших файлов --- actions.dat (вообще все действия), snapshots.json (все снимки, кроме последнего) и lastSnapshot.json.

2. Изменена логика сохранений, теперь действия кампании (ActionCampaign*) также сохраняются, поэтому можно не сохранять снимки при завершении блока кампании (который между ScriptDisableActiveGame и ScriptEnableActiveGame). Вообще теперь снимки сами не создаются (раньше сохранялись через каждые 100 действий), но при загрузке игры может создаться один снимок, в случае если между последним снимком и полученным в результате загрузки снимком более чем 200 действий.

3. Все снимки теперь в json, в бинарном формате остались только действия. Это удобней, +теперь можно визуально сравнивать снимки.

4. !!! Автоматическая генерация методов сериализации с помощью spoon на основе аннотаций и типов полей. Генерируются как методы сериализации в json, так и в бинарный формат.

5. Многочисленные исправления (вспоминается только добавление полей iMin, jMin, iMax, jMax в BaseDrawMain, как минимум это должно увеличить производительность на больших картах, +позволило сделать появление blust'а в DrawCitadelAttack более точным)

+++ app/src/basic/java/ru/ancientempires/BaseThread.java
@@ -2,6 +2,7 @@ package ru.ancientempires;
 
 import android.graphics.Canvas;
 import android.view.SurfaceHolder;
+
 import ru.ancientempires.activity.BaseGameActivity;
 import ru.ancientempires.draws.BaseDrawMain;
 import ru.ancientempires.framework.Debug;
@@ -10,18 +11,18 @@ import ru.ancientempires.framework.MyAssert;
 public class BaseThread extends Thread
 {
 	
-	public static final long	MILLISECONDS_BETWEEN_FRAMES	= 1000 / 30;
-															
-	public BaseGameActivity		activity;
-	public SurfaceHolder		surfaceHolder;
-	public BaseDrawMain			drawMain;
-								
-	volatile public boolean		isRunning					= true;
-															
-	public float				touchY;
-	public float				touchX;
-	public boolean				isTouch;
-								
+	public static final int MILLISECONDS_BETWEEN_FRAMES = 1000 / 30;
+
+	public BaseGameActivity activity;
+	public SurfaceHolder    surfaceHolder;
+	public BaseDrawMain     drawMain;
+
+	volatile public boolean isRunning = true;
+
+	public float   touchY;
+	public float   touchX;
+	public boolean isTouch;
+
 	public BaseThread(BaseGameActivity activity, SurfaceHolder surfaceHolder)
 	{
 		Debug.onCreate(this);
